backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
python-dotenv==1.0.1
websockets==12.0
pydantic==2.6.1
email-validator==2.1.0
python-decouple==3.8